@@ -27,25 +27,24 @@
     ]
   },
   "dependencies": {
-    "d3": "^7.0.3",
-    "prelude-ts": "^1.0.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-resizable-grid-layout": "^0.2.1",
-    "react-scripts": "4.0.3",
-    "rxjs": "^7.3.0",
-    "stream-charts": "^1.0.1"
+    "d3": "^7.9.0",
+    "prelude-ts": "^1.0.6",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-resizable-grid-layout": "file:../react-layout/react-grid-layout/react-resizable-grid-layout-0.3.0-snapshot.tgz",
+    "react-scripts": "^5.0.1",
+    "rxjs": "^7.8.1"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.4.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/d3": "^7.0.0",
-    "@types/jest": "^27.0.2",
-    "@types/node": "^16.9.5",
-    "@types/react": "^17.0.24",
-    "@types/react-dom": "^17.0.9",
-    "typescript": "^4.4.3"
+    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
+    "@types/d3": "^7.4.3",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^22.5.2",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
+    "typescript": "^4.9.5"
   },
   "files": [
     "README.md",
