@@ -17,40 +17,39 @@ import {
 } from 'react-resizable-grid-layout';
 import {lightTheme, Theme} from "./Themes";
 
-// import {Datum, Series, seriesFrom} from "../charts/datumSeries";
-// import {ChartData} from "../charts/chartData";
-// import {regexFilter} from "../charts/regexFilter";
-// import {Chart} from "../charts/Chart";
-// import { defaultMargin } from '../charts/hooks/useChart';
-// import {AxisLocation, defaultLineStyle } from '../charts/axes';
-// import {ContinuousAxis} from "../charts/ContinuousAxis";
-// import {CategoryAxis} from "../charts/CategoryAxis";
-// import {Tracker, TrackerLabelLocation} from "../charts/Tracker";
-// import {Tooltip} from "../charts/Tooltip";
-// import {RasterPlotTooltipContent} from "../charts/RasterPlotTooltipContent";
-// import {formatNumber, formatTime} from '../charts/utils';
-// import {RasterPlot} from "../charts/RasterPlot";
-// import {ContinuousAxisRange} from "../charts/continuousAxisRangeFor";
-import {
-    assignAxes,
-    AxisLocation,
-    CategoryAxis,
-    Chart,
-    ChartData,
-    ContinuousAxis,
-    Datum,
-    defaultLineStyle,
-    defaultMargin,
-    formatNumber, formatTime,
-    RasterPlot,
-    RasterPlotTooltipContent,
-    regexFilter,
-    Series,
-    seriesFrom,
-    Tooltip,
-    Tracker,
-    TrackerLabelLocation
-} from "stream-charts"
+import {Datum, Series, seriesFrom} from "../charts/datumSeries";
+import {ChartData} from "../charts/chartData";
+import {regexFilter} from "../charts/regexFilter";
+import {Chart} from "../charts/Chart";
+import {defaultMargin} from '../charts/hooks/useChart';
+import {AxisLocation, defaultLineStyle} from '../charts/axes';
+import {ContinuousAxis} from "../charts/ContinuousAxis";
+import {CategoryAxis} from "../charts/CategoryAxis";
+import {Tracker, TrackerLabelLocation} from "../charts/Tracker";
+import {Tooltip} from "../charts/Tooltip";
+import {RasterPlotTooltipContent} from "../charts/RasterPlotTooltipContent";
+import {formatNumber, formatTime} from '../charts/utils';
+import {RasterPlot} from "../charts/RasterPlot";
+// import {
+//     AxisLocation,
+//     CategoryAxis,
+//     Chart,
+//     ChartData,
+//     ContinuousAxis,
+//     Datum,
+//     defaultLineStyle,
+//     defaultMargin,
+//     formatNumber,
+//     formatTime,
+//     RasterPlot,
+//     RasterPlotTooltipContent,
+//     regexFilter,
+//     Series,
+//     seriesFrom,
+//     Tooltip,
+//     Tracker,
+//     TrackerLabelLocation
+// } from "stream-charts"
 
 interface Visibility {
     tooltip: boolean;
@@ -111,7 +110,8 @@ export function StreamingRasterChart(props: Props): JSX.Element {
 
     // elapsed time
     const startTimeRef = useRef<number>(new Date().valueOf())
-    const intervalRef = useRef<NodeJS.Timer>()
+    const intervalRef = useRef<NodeJS.Timeout>()
+    // const intervalRef = useRef<NodeJS.Timer>()
     const [elapsed, setElapsed] = useState<number>(0)
 
     // chart time
